% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsdesc.R
\docType{data}
\name{forbes}
\alias{forbes}
\title{Forbes' data on boiling point and  barometric pressure.}
\format{A data frame with 17 rows and 2 variables}
\source{
Applied Linear Regression by Sanford Weisberg
}
\usage{
forbes
}
\description{
Forbes' data giving boiling point (F) and log of barometric pressure (log of inches of mercury) for 17 locations in the Alps and in Scotland. The variables are as follows:
}
\details{
\itemize{
  \item boiling.point. Boiling point in F.
  \item log.pressure. Log of the pressure.
}
}
\keyword{datasets}

