% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsdesc.R, R/rateofreturn.R
\docType{data}
\name{productivity.improvement}
\alias{productivity.improvement}
\title{A dataset containing Measures of Productivity improvement for firms producing electronic computing equipment.}
\format{A data frame with 27 rows and 2 variables}
\source{
Neter Kutner Nachtsheim and Wasserman, Applied Linear Statistical Models, fourth ed., p.703.

Neter Kutner Nachtsheim and Wasserman, Applied Linear Statistical Models, fourth ed., p.703.
}
\usage{
productivity.improvement

productivity.improvement
}
\description{
A dataset containing the rate of return of share of different companies. The variables are as follows:

A dataset containing the rate of return of share of different companies. The variables are as follows:
}
\details{
\itemize{
  \item level. Level of theur average expenditures for researhc and development in the past three years ("Low","Moderate","High")
  \item improvement. Productivity improvement in a scale from 0 to 100.
}

\itemize{
  \item level. Level of theur average expenditures for researhc and development in the past three years ("Low","Moderate","High")
  \item improvement. Productivity improvement in a scale from 0 to 100.
}
}
\keyword{datasets}

