% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsdesc.R, R/rateofreturn.R
\docType{data}
\name{rateofreturn}
\alias{rateofreturn}
\title{Rate of return of share.}
\format{A data frame with 18 rows and 2 variables}
\source{
imaginary

imaginary
}
\usage{
rateofreturn

rateofreturn
}
\description{
A dataset containing the rate of return of share of different companies. The variables are as follows:

A dataset containing the rate of return of share of different companies. The variables are as follows:
}
\details{
\itemize{
  \item Sector. Sector ("Oil" or "Service")
  \item Rate. Rate of return
  ...
}

\itemize{
  \item Sector. Sector ("Oil" or "Service")
  \item Rate. Rate of return
  ...
}
}
\keyword{datasets}

