% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateofreturn.R
\docType{data}
\name{rateofreturn}
\alias{rateofreturn}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 18 rows and 2 variables}
\source{
imaginary
}
\usage{
rateofreturn
}
\description{
A dataset containing the rate of return of share of different companies. The variables are as follows:
}
\details{
\itemize{
  \item Sector. Sector ("Oil" or "Service")
  \item Rate. Rate of return
  ...
}
}
\keyword{datasets}

